Escape format string when filling status buffer.
When running repo status, the process output is given as a format string to
`format' for insertion in the repo status buffer. If repo outputs a %, it won't
be escaped and emacs will complain about missing format arguments. This commit
adds a format string when inserting the text in the status buffer to prevent
this.